pyproject.toml
[project]
name = "techbuddy"
version = "0.1.0"
description = "Local streaming speech transcription over WebSocket"
authors = [
    {name = "sukant-kumar",email = "sukant.kumar@arrise.com"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "fastapi (>=0.135.3,<0.136.0)",
    "uvicorn[standard] (>=0.34.0,<0.35.0)",
    "websockets (>=14.0,<15.0)",
    "pywhispercpp (>=1.2.0,<2.0.0)",
    "loguru (>=0.7.2,<1.0.0)",
    "pyyaml (>=6.0.1,<7.0.0)",
    "streamlit-webrtc (>=0.64.5,<0.65.0)",
    "numpy (>=2.4.4,<3.0.0)",
    "scipy (>=1.17.1,<2.0.0)",
    "faster-whisper (>=1.2.1,<2.0.0)",
    "gunicorn (>=25.3.0,<26.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "openai (>=2.34.0,<3.0.0)",
    "python-dotenv (>=1.2.2,<2.0.0)",
    "tenacity (>=9.1.4,<10.0.0)",
    "minio (>=7.2.20,<8.0.0)",
    "python-multipart (>=0.0.27,<0.0.28)",
]

[project.scripts]
techbuddy-server = "techbuddy.server.app:run"
techbuddy-client = "techbuddy.ui_tk:main"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
ruff = "^0.15.12"

